Added "freshness" to Values
This is a very naive mechanism for doing updates without going fully reactive. A value is fresh if it is fresh and all its dependencies are fresh. It needs calculation if it is not fresh but all its dependencies are.

As we mostly deal with small circuits, this shouldn't be too inefficient